refactor(recovery): extract format-state utility helpers
